+++ moabb/datasets/martinezcagigal2023_checker_cvep.py
@@ -178,6 +178,7 @@ class MartinezCagigal2023Checker(BaseDataset):
             ),
         ),
     )
+    nemar_id = "nm000240"
 
     def __init__(self, conditions=ALL_CONDITIONS, subjects=None, sessions=None, **kwargs):
         deprecated_renames = {"Conditions": "conditions"}