+++ moabb/datasets/ssvep_mamem.py
@@ -326,6 +326,7 @@ class MAMEM1(BaseMAMEM):
            `<https://figshare.com/articles/dataset/MAMEM_EEG_SSVEP_Dataset_I_256_channels_11_subjects_5_frequencies_/2068677?file=3793738>`_
     """
 
+    nemar_id = "nm000119"
     METADATA = DatasetMetadata(
         acquisition=AcquisitionMetadata(
             sampling_rate=250.0,
@@ -828,6 +829,7 @@ class MAMEM2(BaseMAMEM):
            `<https://figshare.com/articles/dataset/MAMEM_EEG_SSVEP_Dataset_II_256_channels_11_subjects_5_frequencies_presented_simultaneously_/3153409?file=4911931>`_
     """
 
+    nemar_id = "nm000120"
     METADATA = DatasetMetadata(
         acquisition=AcquisitionMetadata(
             sampling_rate=250.0,
@@ -1331,6 +1333,7 @@ class MAMEM3(BaseMAMEM):
            `<https://figshare.com/articles/dataset/MAMEM_EEG_SSVEP_Dataset_III_14_channels_11_subjects_5_frequencies_presented_simultaneously_/3413851>`_
     """
 
+    nemar_id = "nm000121"
     METADATA = DatasetMetadata(
         acquisition=AcquisitionMetadata(
             sampling_rate=128.0,