+++ moabb/datasets/thielen2021.py
@@ -134,6 +134,7 @@ class Thielen2021(BaseDataset):
 
     """
 
+    nemar_id = "nm000214"
     METADATA = DatasetMetadata(
         acquisition=AcquisitionMetadata(
             sampling_rate=512.0,